New_backend/AudioImageCarrier-Backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Encryption & Compression
cryptography==41.0.7
zstandard==0.22.0

# Image Processing
Pillow==10.1.0
numpy==1.26.2

# File Operations
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2